small

+++ functions/adversarial.py
@@ -11,10 +11,10 @@ import argparse
 from functions.imagenet_classes import classes
 
 
-def adversarial_examples(image_path, model_name='resnet18', method='Fast Gradient Sign Method',eps = 5, alpha = 1,num_iter = None, show=True):
+def adversarial_examples(image_path, model_name='resnet18', method='Fast Gradient Sign Method',eps = 5, alpha = 1, num_iter = None, show=True):
     
     if num_iter == None:
-        num_iter = round(max(eps+4,eps*1.25))
+        num_iter = round(max(eps+4,eps*1.25)).astype(np.uint16)
         
     if method == 'BI':
         method = 'Basic Iterative'
